telegram-bot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
